feat(core): move storage api code to nodestorage folder
