+++ proxmox/nodes/client.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/bpg/terraform-provider-proxmox/proxmox/api"
 	"github.com/bpg/terraform-provider-proxmox/proxmox/nodes/containers"
+	"github.com/bpg/terraform-provider-proxmox/proxmox/nodes/nodestorage"
 	"github.com/bpg/terraform-provider-proxmox/proxmox/nodes/tasks"
 	"github.com/bpg/terraform-provider-proxmox/proxmox/nodes/vms"
 )
@@ -40,6 +41,14 @@ func (c *Client) VM(vmID int) *vms.Client {
 	return &vms.Client{
 		Client: c,
 		VMID:   vmID,
+	}
+}
+
+// Storage returns a client for managing a specific storage.
+func (c *Client) Storage(storageName string) *nodestorage.Client {
+	return &nodestorage.Client{
+		Client:      c,
+		StorageName: storageName,
 	}
 }
